[refactor] generateDefaultBoardのユーティリティ化
